refactor(linkageRules): extract common function
